Split directives
